Work Dump
Adjusted the function in PhysicsButtonPress() to properly animate the button. Also created a function to allow it to pop back up.
Created an ID system to theoretically allow for multiple Randomizers at once.
Updated the Slots Boss Arena to align better with current planned boss.
modifications to the SlotsController script.
7 new animations for the Slots Boss.